Make response_code long as suggested by CURL documentation
response_code should be a long according to
https://curl.haxx.se/libcurl/c/curl_easy_getinfo.html. Before this
commit it was a std::int32_t. In g++ this caused the call to
curl_easy_getinfo to overwrite curl_error and set it to 0 which made the
BadHost tests fail.

+++ cpr/session.cpp
@@ -352,7 +352,7 @@ Response Session::Impl::makeRequest(CURL* curl) {
     auto curl_error = curl_easy_perform(curl);
 
     char* raw_url;
-    std::int32_t response_code;
+    std::int64_t response_code;
     double elapsed;
     curl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &response_code);
     curl_easy_getinfo(curl, CURLINFO_TOTAL_TIME, &elapsed);
