+++ include/cpr/response.h
@@ -16,14 +16,14 @@ class Response {
 
     template <typename TextType, typename HeaderType, typename UrlType, typename CookiesType,
               typename ErrorType>
-    Response(const std::int32_t& p_status_code, TextType&& p_text, HeaderType&& p_header, UrlType&& p_url,
+    Response(const std::int64_t& p_status_code, TextType&& p_text, HeaderType&& p_header, UrlType&& p_url,
              const double& p_elapsed, CookiesType&& p_cookies = Cookies{},
              ErrorType&& p_error = Error{})
             : status_code{p_status_code}, text{CPR_FWD(p_text)}, header{CPR_FWD(p_header)},
               url{CPR_FWD(p_url)}, elapsed{p_elapsed}, cookies{CPR_FWD(p_cookies)},
               error{CPR_FWD(p_error)} {}
 
-    std::int32_t status_code;
+    std::int64_t status_code;
     std::string text;
     Header header;
     Url url;
